-> Convert cohort_files and count_public_data_types to BQ w/params
